Only use multiprocessing.Queue workaround on OSX, rename it from bugfix_mp for aesthetics
